+++ packages/authentication/modules/user/index.ts
@@ -1,2 +1,3 @@
-export { default as useUser } from './useUser'
+export { default as getUser } from './getUser'
 export { default as useSimpleTokenUser } from './useSimpleTokenUser'
+export { default as useUser } from './useUser'