+++ packages/test/utils/mocks.ts
@@ -3,7 +3,7 @@ import Cookies from 'js-cookie'
 
 jest.mock('js-cookie')
 
-export type CookiesGetByNameFn = (name: string) => string
+export type CookiesGetByNameFn = (name: string) => string | undefined
 
 export const cookiesMock = jest.mocked(Cookies)