+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @baseapp-frontend/graphql
 
+## 1.0.2
+
+### Patch Changes
+
+- Imports the `getToken` function from the `@baseapp-frontend/utils` package.
+- Removes `next` dependency.
+- Updated dependencies
+  - @baseapp-frontend/utils@1.4.0
+
 ## 1.0.1
 
 ### Patch Changes