+++ packages/utils/CHANGELOG.md
@@ -1,10 +1,17 @@
 # @baseapp-frontend/utils
 
+## 1.4.0
+
+### Minor Changes
+
+- Creates a `getToken` function that can get tokens from the server and client sides.
+- Adds `next` dependency.
+
 ## 1.3.7
 
 ### Patch Changes
 
-- Refactor `setFormApiErrors` so it only sets the form's errors if the API response is appropriate and if the field matches the form key. 
+- Refactor `setFormApiErrors` so it only sets the form's errors if the API response is appropriate and if the field matches the form key.
 
 ## 1.3.6